create only one tiscaf HApp that is extensible
in tiscaf, sessions are managed per `HApp`, thus there must be only one
such application living in the \BlueLaTeX server, otherwise we would
have several instances of the same session (one per modular
application).
The new extensible application design has only one tiscaf `HApp` and
modular Rest interfaces are registered in this single application which
manages the sessions and the path resolving process.
This is transparent for the module implementors