forward the join/part messages to the actor
this allows the resource actor to maintain a list of connected users.
It is assured by the dispatcher that a join message is sent only if the
user was not already using this resource. In the same spirit, the part
message is only sent to the actor if the user joined this resource
before and never left