change the RestApi to make it easier to use
the old library for matching against some URL and request data was full
of strange code, and almost enforced the user to pattern match against a
list.
Now a RestApi gets a request as input data and there are some extractor
that may be used (or not) be Rest Api developers to pattern match
against the path and/or query parts.

For example, one can write:
  POST {
    case path"$paperId/compile" =>
      ???
    case req @ path"$paperId/compile" ? query"a=${int(p1)}&b=$p2" =>
      ???
    case req @ path"$paperId/q" =>
      ???
    case req =>
      val path = req.uriPath
      val query = req.query
      val accepts = req.header("Accepts")
      ???
  }

As can be seen in the last example, does not use the string
interpolators or extractors at all, everybody is free to redefine them
as they wish.

The interpolators and extractors provided by the RestApi trait are:
 - `long` that extracts a string to a long
 - `int` that extracts a string to an int
 - `date` that extracts a string with format
   "yyyy-MM-dd'T'HH:mm:ss.SSSZ" to a Date object
 - `?` that extract an `HReqData` to a pair of string representing
   respectively the path and the (non empty) query string
 - `path` the interpolator extractor that allows people to extract path
   fragments from either a string or an `HReqData?`
 - `query` the interpolator extractor that allows people to extract
   query parameters from the query string