merge common and http module in a core module
let's be realistic, what are the core features of a \BlueLaTeX server? A
synchronization server providing an HTTP Api.
Once this is stated and admitted by everybody, we can answer this new
question: why should we separate the HTTP module from the common module?
The answer is quite simple: No reason to that.

The 'look, we separate responsibilities and features in a really clean
way' argument does not hold very long when confronted to real-world
application maintenability.

This is symptomatic of this disease called 'overmodularization' that I
got when starting the new \BlueLaTeX version with OSGi. However I treated
myself and I am well again. It is never too late to fix problems and
thus both modules are now merged into one 'core' module which name is
really explicit. If you want to extend \BlueLaTeX, we build a module
using core features, that's it.

I am pretty sure future maintainers/developers will thank me for that
decision to reduce the number of modules to its strict minimum.

  Ikea / furnishing = OSGi / software: They allow to have very modular
  things, but trying to put them together generally drives you crazy
  -- Mario Fusco