fix OSGi problems
Using OSGi is great... Once you spent hours to make all your libraries
work in OSGi container. So far, two libraries were having troubles:
 - javax.mail
 - scalate

What "funny" is, is that they are packaged as OSGi bundles by their
respective packager, so I thought naively "Hey! That's great! Just drop
this jar into my bundle repository!"
I was *so* wrong and didn't realized it by then.

scalate works great in OSGi, but only if you set the current classloader
to the classloader that loaded the scalate classes before any call to
the API. In our case that's easy because the template engine is wrapped
and should never be called directly. There is a single point where it is
used.

So far, scalate: Check!

What about javax.mail?
This one is funny, it depends on activation.jar, which does seem to be
an OSGi bundle. No problem I wrote my wrapper for this case, using BND,
that just exports every package in the bundle.
Is that it? You wish!
The funny stuff is that javax.mail and activation.jar must have the same
classloader because they are so tied together that one can wonder why on
earth they are in two different jar files!
The solutions I found browsing the web were bad, really bad, so bad that
I almost had to throw up. I didn't want to merge both jar into one,
because let be honest, who wants to complicate his workflow to handle
this packaging non-sense?
I don't at least.

And then I remembered that I used to be an OSGi guru (well, I like to
think I once was) and that there are these underestimated stuffs called
fragments! In my case I know that only javax.mail uses activation and
that I'd like the latter to be part of the former, in particular it
would be great if they had the same classloader.
Tadam! Thank you fragments! That's what they were made for (at least
that's how I understand them): you have a bundle that is entirely part
of another, that cannot leave without it and should be part of it but
for some (good or bad) reason could not be include in it. This fragment
will get the host classloader and everybody is happy, I can send emails!

If it were not my idea, I'd so much kill the guy who said "Hey guy,
let's use OSGi for \BlueLaTeX, after all, we have modules sharing
services, that's a good use case!"

Everything to work correctly in OSGi now, but wait, I didn't use any
actor yet, there is still some possibility to have a complex system to
make work in OSGi (even if once more, akka is packaged as an OSGi
bundle, who knows, things can (and will with high probability) go wrong.