[imp #2] Fixes & update to last protocol version
* Storage manager is now common: it is created in dispatcher
instead of being instancied in every actors.

* Synchronization handles multiple files:
Until now, a synchronization actor could only handle one file.
It is now possible for an actor to deal with all synchronized resources given
a paper Id. Each document is to be carried in the synchronization message.

* Add some log traces in SyncActor

* Add exception handling on DMP method invocation:
A call to the DMP.diff_fromDelta() method would raise an
InvalidArgumentException if the given delta length
does not match the shadow length.
Fix this by enclosing this call in a Try construction.

* Add PersistPaper message to SyncActor:
Uppon reception of a PersistPaper message, the synchronization actor will save
the content of the in-memory document ('server' representation of the document)
to the disk.

* Add unit tests