extract configuration loader to own bundle
All bundles potentially use the configuration loader, and to avoid
cyclic dependencies, the best way to do it is to have a bundle whose
single responsibility is to register the configuration loader.
Thus, it will probably be the first bundle that is started and then
everybody is happy and can use it.