delete references to unregistered user
If a user is single author for any paper, then he cannot unregister. He
must first either delete the paper data or give another user the author
role for this paper.
If everthing went well, delete references to user as author or reviewer
for all papers he plays a role in.