Add logger to BlueLet
A request handler takes a logger to keep track of events that happened
during request processing.
The next step is now to actually log stuffs everywhere where there is a
TODO (and probably a lot more other places) and to add standrad logging
configuration that logs into files.

What would be interesting is to have something like `PaperLoggin` which
logs stuffs in a paper specific logger (e.g. in a specific log file).
Maybe the SiftingAppender class of logback can be used to implement this
feature in the slf4j way.
We must also see how this can integrate with the OSGi log service.