Move actor system to common bundle
The actor system may be used by a lot of bundles, so this is a common
stuff. Actually the more I am thinking about it, the more I want to
register everything but the core http api in the common activator.

The core bundle is just the core features that allow \BlueLaTeX to
manage people, sessions and papers. Nobody depends on it directly,
everybody needs the common features.

And if somebody needs this, lets just declare it a dependency when
needed. But this was of organizing stuffs looks better to me.