Refactoring: use custom UserInfo type
For the moment the `common.UserInfo` type is just an alias for the
`sohva.UserInfo` type. But this may differ in the future. It does not
cost anything to use an alias here in this case, and make future
changes easier (when the code base will be much much larger).