New test scenarios
We now test the saving of user settings, when everything goes well. This
scenario allowed me to improve the test client to also return the
headers (needed to get the `ETag` header when getting the data), and I
could fix some problems that made it impossible to save user settings.