Fix email encoding
Sending Email in utf-8 was not supported. We are in 2014, and it is a
shame that this is not the default, but we have to live with this. It is
also highly probable that people registering to such a system will have
names containing non-ASCII characters. As it is always better when it is
explicit, setting the email content encoding to utf-8 avoids big
troubles later in production.