[imp #2] Implementation of Broadcast messages
Each connected user gets a mailbox where the broadcasted messages issued
from other users are stored. When a user wants to retrieve its messages, a flag
is set in the protocol, and the answer contains the list of the user's messages.

As a side note, the original mobwrite protocol is more simple concerning
broadcast message: a client can only send one broadcast message (which can be
overwritten or blanked by future broadcast message), and it is only associated
with a file, not with a (file,revision) tuple.