Move common stuffs to common module
Finally! I took some time to refactor and rearrange the module
organization. It was not as straight forward as I originally thought,
but mainly because of my stupidity, not to log the configuration loader
issue I encounter every time I move anything related to configuration.
I am not happy with the way it is done for the moment, and it does not
seem to cope with OSGi environments and hierarchical loading of
configuration. I should investigate further this issue, but this time it
should not change a lot anymore from the modules point of view.

Anyway, test scenarios all pass again! And the different modules are
much more consistent:
 - `blue-common` really does contain common stuffs and services (http
   server, couchdb stuffs, ...)
 - `blue-core` really does contain core features only, which include,
   but are not limited to: user and session management, paper
   management.

No cyclic dependencies anymore, nothing weird in the design anymore! I'd
love to think that it's a good design now, but I tend to think I will
blame myself in a few weeks on how stupid I was back then, not to have
seen basic problems I will for sure encounter really soon. If this does
not happen, for any reason, then I really should check if I did not sign
any pact with Cthulhu to make it work, desperate as I was to make it
finally work. Who will live, will know!