Unify resource names
We have no use of file extensions to access the \BlueLaTeX resources, it
is much cleaner to have uniform resource names, also for generated pdf
files and all other pdf stuffs. Same applies for the zip archive that
can be downloaded.

This work prepares for the API extension required by #18.