Remove legacy parameter
The `density` parameter does not cope well with the SyncTeX feature.
Besides, below 100 (or 96 maybe) the rendered result is barely readable,
above, it is really huge. This parameter raises some other probems (see
tickets #15 and #19).

What could be useful however, is a parameter `thumbnail` that would
generate a small version of the page for an overview. But even that I am
not sure it would be useful to have...