[imp #23] fix last modification date interface
The server was returning the lastest modification date for all its associated
document.
However, if there was no document opened, this lead to an invalid call to method
List.last().

Now, the last modification date is also stored in the actor itself (we still
keep per document date, as they might be useful), and updated accordingly.