Migrate to new sohva version
This is no simple commit, as sohva changes its Http library in version
1.0.0, and now uses spray and akka. This implies that an `ActorSystem`
must be present when it is instantiated.

Making this change, I discovered that we had a terrible way to deal with
sohva instances and that we actually instantiated an awful lot of them
where only one is actually needed and recommended. That's the reason why
this patch looks quite huge. All `HLet` gained a new parameter which is
the sohva client they can use, instead of creating a new one for each
new `HLet`. No new feature was added, and it is much of a cosmetic
change.

This change also permitted to upgrade the akka version we are using,
which upgraded the version of config and made some deprecation warning
appear. Duration in configuration are now properly handled with the more
flexible `Config.getDuration` method.

Thanks to this change and cleanup , \BlueLaTeX should scale better now.