Allow parsing of large chunks of text in EditCommand
Change the regex used in the Add command.
Do not use the valid regex anymore as Java RE is not able to apply it
on more than a few characters, and fallback to a quickest, insecure version
instead...
This *MUST* be reworked in a later version.

This fix is part of #30