Extract paper and class from TeX source
Instead of reading the paper title from the PDF metadata, which are not
set by default by the pdflatex compiler, this piece of information is
extracted from the source.
This is not really reliable as is because now that there are support for
several files, the title and the class may be defined in an other than
the main paper file, but it should be ok in most of the cases.