[fix #51] Correctly return default index file
When the base URL is accessed with no file name, an empty file was
returned. This is due to how the `getResource` method works with
resources in a jar file (actually it would behave the same in any zip
file). If the jar file contains directory entries, an empty stream to
the directory entry is returned, and the `WebLet` think the resource is
found. It is probably safe in \BlueLaTeX to assume that the web client
does not serve any empty static file. Thus, we can filter them out.

Closes #51