Code improvements
While debuggin the problem with the join message I had to add some
logging here and there, and I fixed some code style issues that led to
timeouts under certain circumstances. It would be a pity to throw
everything away, so I include it here. The code should be more robust
now.