[imp #88] Add paper name to identify it
When a user creates a paper, he must give a name that is used in paper
list for further reference to it.

Using a name that is not the title extracted from the source is better
for several reasons:
 - Using a simple regular expression to extract the title from the
   source file is not a really robust method,
 - Source title may contain control sequences that are not really user
   friendly to display,
 - Several papers may have similar titles which will make it more
   difficult to distinguish between them.

For all these reasons we found it preferable to ask for a name that the
user can freely chose and edit.