[imp #42] Synchronization actor code clean-up
Lots of clean-up done in this commit:
 * The actor system now uses immutable parameters. This should allow more
advance and fancy akka-stuff, such as actors over network, etc. To do this, all
mutable parameters have been encapsulated in an immutable case class, and the
code now makes use of the 'context.become()' pattern. There is still use of
mutables variable in the code, but far less than before, and they should not
interfere with akka possibilities.
 * Separate code of dispatcher and actor. It is much easier to maintain this
way. By the way, correctly rename 'SynchroDispatcher' to 'SyncDispatcher'.
 * Rename unit test from 'SynchroDispatcher' to 'SyncActor', since it actually
tests the actor and not the dispatcher (+ it adds naming consistency).
 * Add of a package file to define type aliases for PaperId, PeerId, etc. In
fact, it may be a good move to actually define these in the blue-core package.

This commit is quite big and should have been done in multiple, smaller commits.
For that, I apologize. No kittens were harmed in the making of this patch.