[imp #94] Add new system compilers
Adding common \LaTeX compilers makes it possible to compiler more
documents.
So, by default we now have:
 - pdflatex (default compiler)
 - latex + dvipdfm,
 - xelatex,
 - lualatex

\BlueLaTeX assumes that these tools are installed on the system and
does not check for their existence.

If problems arise on some platforms, we should have a look at how to
make it detect if these tools are present in a portable way (well,
problem is probably on windows, which has no `which` command, and I am
not aware of any similar command, help is welcome here).

This commit closes #94