[imp #55] Improve configuration loading
The default values are now in a file named `application.conf` so that
keys are resolved with the local overrides for configuration. This makes
it possible to simplify configuration overrides when substitutions are
used. The keys using substitution in the default configuration must not
be overridden as well.

By making sure that the local files are loaded at last, we now that
their values will be taken over the default ones, as specified in the
HOCON file format.