[imp #112] Make compilation type configurable
Server administrator can choose whether papers on this \BlueLaTeX
instance are compiled explicitly by the users or as a periodic
background task.
The default is now the more intuitive explicit compilation type.

The configuration sent back to the client was extended to inform whether
it should automatically resend a compilation request when the previous
one is over.

Fixes #112