Make lift-json version OSGi friendly again
OSGi bundles require to have a version number following a strict scheme
which lift-json doesn't have.

I know I'm guilty about having pushed this without checking this part,
and I am very sorry for that. It will never happen again... Or maybe it
will, but bumping version shouldn't lead to this, if we only use OSGi
friendly libraries, which is something we are going towards.