[fix #134] Remind username and email
When a user registers to \BlueLaTeX, a validation email is sent with
instructions to set the password.

Reminding about the chosen username and email address might be good even
if the user receives the mail at the registered address, and the
username appears in the validation link.

Explicit stuffs are always better than implicit ones.

Closes #134