[fix #156] Fix regression in paper option
The fix is quite simple, the problem is real however. The role saving
function must be actually executed for the roles to be persisted into
the database.

This bug only impacts the paper options screen, contributors can still
be added or removed by using the new "Share" menu in the paper editor.