[imp #173] Add button to cleanup files
Sometimes, we need to explicitly clean files in the compilation
directory. The server has supported this for a long time now, without
being exposed in the web client.
That is now the case.

I also used this chance to improve the recompilation after a cleanup. If
the .aux file cannot be found, the compilation is performed, even if the
paper did not change. This makes sense, because if the user cleaned up
the compilation directory, it probably means that he wants a fresh
compilation to be performed.