Remove dependency to scalate
scalate is really good as a template engine for things like jsp or such.
In our case, we do not generate html pages dynamically, only paper
templates and emails are loaded as templates.
Using our own lightweight mustache template engine we avoid having a
resident scala compiler, we avoid clsasloader problems with OSGi, the
warmup phase is much smaller, and it was fun and easy to implement.

Moreover this makes less dependencies, once we do not use lift-json
anymore, no dependency on the scala compiler will be required anymore,
and that will be a great step.