[fix #214] Ignore messages from unknown peers
If a peer sends a message to broadcast but is still unknown to the
server, his message is simply ignored.
Before the first synchronization, a peer cannot send his position but
I'd say it is safer this way anyway and is not much of a problem.

In any way, the server should not throw any exception in this case.