[fix #209] Always write files in UTF-8
Since the `FileWriter` class selects the JVM default encoding, if the
underlying platform is not using UTF-8 we have problem loading the
files because we explicitly want UTF-8 encoding.

The Java 6 way of doing it is not that direct, but is not complicated
either.
Beware that no BOM is written with this technic, which should not be a
problem for us, because our reading methods do not expect BOM either.

In the next version we can use scalax.io or some Java 7 methods, if we
want to drop the Java 6 support.